version 0.0.2

+++ Guardian del Colgante/Assets/Scripts/Jugador/pebeteScript.cs
@@ -35,7 +35,7 @@ public class PebeteScript : MonoBehaviour
         }
     }
 
-    public void ConfigurarDamage(int nuevoDamage)
+    public void ConfigurarDamage(float nuevoDamage)
     {
         dano = nuevoDamage;
     }
